+++ src/headless/plugins/chat/index.js
@@ -38,6 +38,8 @@ converse.plugins.add('converse-chat', {
             'auto_join_private_chats': [],
             'clear_messages_on_reconnection': false,
             'filter_by_resource': false,
+            'prune_messages_above': undefined,
+            'pruning_behavior': 'unscrolled',
             'send_chat_markers': ["received", "displayed", "acknowledged"],
             'send_chat_state_notifications': true,
         });